backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
requests==2.31.0
psycopg[binary]==3.2.13
python-dotenv==1.0.0
gunicorn==21.2.0
